feat: preform watch comms first then local state
This better address the _async_ nature of the
watch communication. We can now handle failure
communication cases better.

This also:
- Removes the messy higher order feeding communication reducer
- Sets up for receiving feeding state changes from the main app!

In the future, having the watch communication logic move out of the
individual views would be nice, but sending async actions is difficult
to handle right now.
