+++ Watch Extension/HomeView.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct HomeView: View {
     @ObservedObject var store: Store<AppState, AppAction> = {
         let s = Store(initialValue: AppState(),
-                      reducer: higherOrderFeedingCommunicationReducer(appReducer))
+                      reducer: appReducer)
         TimerPulse.shared.store = s
         TimerPulse.shared.start()
